Fix crash: race condition
Launchpad putton pushes are handled in their own thread, so when a snapshot is
requested via the Launchpad the program change is actioned upon by the main
GeoLEDic thread while we're still using the program, essentially cutting the
branch we're sitting on.

Use a mutex to prevent that from happening. This is not really a nice solution,
but time's not on our side, and it works.