LaunchPad: add takeover mode
When a fader is touched on the LaunchPad, the LaunchPad takes
over control for the associated CC, causing messages at the input
port to be dropped.

So for instance, if Ableton is changing a CC, it loses control as soon
as there's manual interaction with the same CC on the LaunchPad.

To show that take over is active for a given fader, the button below it
(the one with the line) is lit on the LaunchPad Pro MK3. If that button is then
pushed, control is returned to the input.

There's currently no way to return control on the LaunchPad Mini, other than via
a program change.